+++ tests/e2e/standard/test_init.py
@@ -117,7 +117,7 @@ class TestInitRun(BaseE2ETest):
 
         repo.git.add("some_file.txt")
 
-        assert repo.is_dirty()
+        assert repo.is_dirty(index=False)
         with neptune.init_run(project=environment.project) as run:
             run.sync()
             assert run.exists("source_code/diff")